requirements.txt
# 美国白蛾预测项目依赖包

# 核心依赖
numpy>=1.20.0
pandas>=1.2.0
scikit-learn>=0.24.0
torch>=1.8.0
torchvision>=0.9.0

# 地理空间数据处理
geopandas>=0.9.0
rasterio>=1.2.0
shapely>=1.7.0
fiona>=1.8.0
pyproj>=3.0.0

# 数据处理和分析
openpyxl>=3.0.0
xlrd>=2.0.0
scipy>=1.6.0
joblib>=1.0.0

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0
plotly>=5.0.0
folium>=0.12.0

# 深度学习增强
tqdm>=4.60.0
tensorboard>=2.5.0

# 配置和日志
pyyaml>=5.4.0
python-dotenv>=0.19.0

# 开发和测试
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# 其他工具
click>=8.0.0
pathlib2>=2.3.0